+++ src/model/shapes/CompoundShape.java
@@ -23,7 +23,7 @@ public class CompoundShape implements IShape {
   }
 
   @Override
-  public void place(int x, int y) {
+  public void place(double x, double y) {
     this.posn.set(x, y);
   }
